+++ Source/HtmlRenderer/Core/Handlers/BordersDrawHandler.cs
@@ -126,16 +126,16 @@ namespace HtmlRenderer.Core.Handlers
                     switch (border)
                     {
                         case Border.Top:
-                            g.DrawLine(pen, (double)Math.Ceiling(rect.Left), rect.Top + box.ActualBorderTopWidth / 2, rect.Right - 1, rect.Top + box.ActualBorderTopWidth / 2);
+                            g.DrawLine(pen, Math.Ceiling(rect.Left), rect.Top + box.ActualBorderTopWidth / 2, rect.Right - 1, rect.Top + box.ActualBorderTopWidth / 2);
                             break;
                         case Border.Left:
-                            g.DrawLine(pen, rect.Left + box.ActualBorderLeftWidth / 2, (double)Math.Ceiling(rect.Top), rect.Left + box.ActualBorderLeftWidth / 2, (double)Math.Floor(rect.Bottom));
+                            g.DrawLine(pen, rect.Left + box.ActualBorderLeftWidth / 2, Math.Ceiling(rect.Top), rect.Left + box.ActualBorderLeftWidth / 2, Math.Floor(rect.Bottom));
                             break;
                         case Border.Bottom:
-                            g.DrawLine(pen, (double)Math.Ceiling(rect.Left), rect.Bottom - box.ActualBorderBottomWidth / 2, rect.Right - 1, rect.Bottom - box.ActualBorderBottomWidth / 2);
+                            g.DrawLine(pen, Math.Ceiling(rect.Left), rect.Bottom - box.ActualBorderBottomWidth / 2, rect.Right - 1, rect.Bottom - box.ActualBorderBottomWidth / 2);
                             break;
                         case Border.Right:
-                            g.DrawLine(pen, rect.Right - box.ActualBorderRightWidth / 2, (double)Math.Ceiling(rect.Top), rect.Right - box.ActualBorderRightWidth / 2, (double)Math.Floor(rect.Bottom));
+                            g.DrawLine(pen, rect.Right - box.ActualBorderRightWidth / 2, Math.Ceiling(rect.Top), rect.Right - box.ActualBorderRightWidth / 2, Math.Floor(rect.Bottom));
                             break;
                     }
                 }
